money-tracker/backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
pymysql==1.1.0
pyjwt==2.8.0
werkzeug==2.3.7
flask-cors==4.0.0
cryptography==41.0.7
sqlalchemy==2.0.23
